serial: msm: Add support for halved sampling rate
Newer SoCs (employing GENI SE QUP v2.5 and newer), determinate by reading
a register of the QUP wrapper, so the parent device of the UART that this
driver handles require a halved sampling rate. Failing to account for it
doubles the effective baud rate or breaks UART completely.

Add the necessary infra and missing clocks to take care of it. While at
it, fix up some glaring style errors.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>